Add second TDD Case:
Find 0 Recomentations by name

Given 0 recommendations it should return an Error.
The next step would be check which kind of exception foursquare returns from this case. Catch this exception and convert it to a customised exception with 400.
But this involves creating the Foursquare Client and check which exception is returning and implement catch.
A TODO was added to refactor this case after the Foursquare API was built.